+++ frontend/src/routes/index.js
@@ -10,7 +10,6 @@ import Login from "../pages/Login";
 import Connections from "../pages/Connections";
 import SettingsCustom from "../pages/SettingsCustom";
 import Financeiro from "../pages/Financeiro";
-import SupervisorPanel from "../pages/SupervisorPanel";
 import RH from "../pages/RH";
 import PublicJobBoard from "../pages/PublicJobBoard";
 import Users from "../pages/Users";
@@ -92,7 +91,6 @@ const Routes = () => {
               <LoggedInLayout>
                 <Route exact path="/financeiro" component={Financeiro} isPrivate />
                 <Route exact path="/rh" component={RH} isPrivate />
-                <Route exact path="/painel-vigia" component={SupervisorPanel} isPrivate />
 
                 <Route exact path="/companies" component={Companies} isPrivate />
                 <Route exact path="/" isPrivate render={(props) => (